+++ NEWS.md
@@ -1,6 +1,6 @@
 # hydroscoper 0.1.0.9000
 
-* Follow hydroscope's multi-page front end
+* Support hydroscope's multi-page front end
 
 # hydroscoper 0.1.0